Squash alembic revisions
